+++ cloudy_salesforce/client/salesforceclient.py
@@ -211,15 +211,14 @@ class SalesforceClient:
         body: dict | None = None,
         params: dict | None = None,
     ) -> dict[str, Any] | list[dict[str, Any]]:
-        if url.startswith("http://") or url.startswith("https://"):
-            request_url = url
-        else:
-            request_url = f"{self.get_instance_url()}{url}"
-
         reauthenticated = False
         rate_limit_attempt = 0
 
         while True:
+            if url.startswith("http://") or url.startswith("https://"):
+                request_url = url
+            else:
+                request_url = f"{self.get_instance_url()}{url}"
             try:
                 response = self.get_session().request(
                     method,